GMDX-393-android-create-sample-repo-from-scracth - code cleaning
