+++ GCMessengerSDKSample/app/src/main/java/com/genesys/gcmessengersdksample/data/SampleData.kt
@@ -1,0 +1,17 @@
+package com.genesys.gcmessengersdksample.data
+
+import com.google.gson.JsonObject
+
+data class SampleData(
+
+    /**
+     * The Account (Bot/Bold/Async)
+     */
+    val account: JsonObject?,
+
+    /**
+     * true if there was a restoration request
+     */
+    val restoreRequest: Boolean
+
+)
