+++ GCMessengerSDKSample/app/src/main/java/com/genesys/gcmessengersdksample/data/defs/FieldType.java
@@ -1,0 +1,19 @@
+package com.genesys.gcmessengersdksample.data.defs;
+
+
+import androidx.annotation.StringDef;
+
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+
+@Retention(RetentionPolicy.SOURCE)
+@StringDef({FieldType.Title, FieldType.TextInput, FieldType.Switch, FieldType.ContextBlock, FieldType.Options, FieldType.Option})
+
+public @interface FieldType {
+    String Title = "title";
+    String TextInput = "textInput";
+    String Switch = "switch";
+    String Options = "radioGroup";
+    String Option = "radioButton";
+    String ContextBlock = "ContextBlock";
+}
