+++ GCMessengerSDKSample/app/src/main/res/layout/activity_main.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/white"
-    tools:context="com.genesys.mobile.messenger.sdk.gcmessengersdksample.MainActivity">
+    tools:context="com.genesys.gcmessengersdksample.MainActivity">
 
     <FrameLayout
         android:id="@+id/content_main"
